Lots of stuff
Okay, so I'm still having an issue with file output I'll need to fix
eventually if i want to display a high score or something. The good news
is I finally did the obvious thing to fix my Tetro recycling problem I
was originally using it to look into.

I'm playing around with the scoring system now, and may look into
retrying the queue system in the near future.