yet another debug failure
I'm trying to create a second array that I can use to create a second
copy of the grid right before it gets messed up. Ideally once the
array's done mysteriously screwing itelf up I can just copy the perfect
backup over it.
For reasons I have yet to completely sort out, the compiler is throwign
up conversion issues.

Something else of note for anyone who seriously wants to try to help
with this mess:
Whenever a position in the bottom row of the array is occupied, it
somehow fills in the poition in the top row one slot to the left.
messing with this top row also changes the bottom row in turn. I haven't
got a clue why this is happenning, and if I could eliminate that it'd
kill all of my current problems.

Also, if anyone else is seriously looking at this, I flip the array
upside down, so what I consider the 'top" is actually the
highest-numbered row.